Move last remaining analyzers to their own module
